requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djoser>=2.2.0
mysqlclient>=2.2.0
django-cors-headers>=4.3.0
